+++ src/TextShow/Data/Complex.hs
@@ -1,5 +1,10 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
+#if __GLASGOW_HASKELL__ == 800
+-- See Note [Increased simpl-tick-factor on old GHCs]
+{-# OPTIONS_GHC -fsimpl-tick-factor=200 #-}
+#endif
 
 {-|
 Module:      TextShow.Data.Ratio
@@ -30,3 +35,22 @@ instance TextShow a => TextShow (Complex a) where
 
 -- | /Since: 2/
 $(deriveTextShow1 ''Complex)
+
+{-
+Note [Increased simpl-tick-factor on old GHCs]
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Compiling certain text-show modules with optimizations on old versions of GHC
+(particularly 8.0 and 8.2) will trigger "Simplifier ticks exhausted" panics.
+To make things worse, this sometimes depends on whether a certain version of
+the text library is being used. There are two possible ways to work around
+this issue:
+
+1. Figure out which uses of the INLINE pragma in text-show are responsible
+   and remove them.
+2. Just increase the tick limit.
+
+Since executing on (1) will require a lot of effort to fix an issue that only
+happens on old versions of GHC, I've opted for the simple solution of (2) for
+now. Issue #51 is a reminder to revisit this choice.
+-}