+++ tests/Instances/GHC/RTS/Flags.hs
@@ -8,12 +8,8 @@
 {-# LANGUAGE TypeSynonymInstances #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 # if __GLASGOW_HASKELL__ == 802
--- For whatever reason, compiling this module with optimizations on GHC 8.2.2
--- triggers a "Simplifier ticks exhausted" panic, and increasing the tick limit
--- doesn't seem to help. Moreover, 8.2.2 is the /only/ GHC version I have seen
--- this happen on. Life is short, so I'm just going to work around this by
--- disabling optimizations on 8.2.2 only.
-{-# OPTIONS_GHC -O0               #-}
+-- See Note [Increased simpl-tick-factor on old GHCs] in TextShow.Data.Complex
+{-# OPTIONS_GHC -fsimpl-tick-factor=200 #-}
 # endif
 #endif