+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### next [????.??.??]
+* Work around a GHC 8.0–specific issue in which GHC's simplifier ticks would
+  become exhausted, causing compilation to fail.
+
 ### 3.9.4 [2021.12.26]
 * Allow the test suite to build with `text-2.0.*` and
   `transformers-compat-0.7.1`.